feat(income-plan): Move shared messages, types, constants and utils to core (#18874)
* [TS-1040] Use messages, types, constants and utils in core

* Move defaultIncomeTypes to constants

* Update text
