+++ libs/application/templates/social-insurance-administration/core/src/lib/constants.ts
@@ -135,3 +135,56 @@ export const maritalStatuses: {
   '0': married,
   L: married,
 }
+
+export const FOREIGN_BASIC_PENSION = 'Erlendur grunnlífeyrir'
+export const FOREIGN_PENSION = 'Erlendur lífeyrir'
+export const FOREIGN_INCOME = 'Erlendar tekjur'
+export const INTEREST_ON_DEPOSITS_IN_FOREIGN_BANKS =
+  'Vextir af innstæðum í erlendum bönkum'
+export const DIVIDENDS_IN_FOREIGN_BANKS =
+  'Arður af hlutabréfa eign í erlendum bönkum'
+export const ISK = 'IKR'
+export const INCOME = 'Atvinnutekjur'
+
+export enum RatioType {
+  YEARLY = 'yearly',
+  MONTHLY = 'monthly',
+}
+
+export const defaultIncomeTypes = [
+  {
+    income: 'yearly',
+    currency: 'IKR',
+    incomeType: 'Lífeyrissjóður',
+    incomePerYear: '0',
+    incomeCategory: 'Lífeyrissjóðstekjur',
+  },
+  {
+    income: 'yearly',
+    currency: 'IKR',
+    incomeType: 'Laun',
+    incomePerYear: '0',
+    incomeCategory: 'Atvinnutekjur',
+  },
+  {
+    income: 'yearly',
+    currency: 'IKR',
+    incomeType: 'Vextir af innistæðum',
+    incomePerYear: '0',
+    incomeCategory: 'Fjármagnstekjur',
+  },
+  {
+    income: 'yearly',
+    currency: 'EUR',
+    incomeType: 'Erlendur lífeyrir',
+    incomePerYear: '0',
+    incomeCategory: 'Lífeyrissjóðstekjur',
+  },
+  {
+    income: 'yearly',
+    currency: 'IKR',
+    incomeType: 'Vextir af verðbréfum',
+    incomePerYear: '0',
+    incomeCategory: 'Fjármagnstekjur',
+  },
+]
