+++ libs/application/templates/social-insurance-administration/income-plan/src/forms/IncomePlanForm.ts
@@ -11,8 +11,23 @@ import {
   buildTableRepeaterField,
 } from '@island.is/application/core'
 import Logo from '@island.is/application/templates/social-insurance-administration-core/assets/Logo'
+import {
+  DIVIDENDS_IN_FOREIGN_BANKS,
+  FOREIGN_BASIC_PENSION,
+  FOREIGN_INCOME,
+  FOREIGN_PENSION,
+  INCOME,
+  INTEREST_ON_DEPOSITS_IN_FOREIGN_BANKS,
+  ISK,
+  MONTHS,
+  RatioType,
+} from '@island.is/application/templates/social-insurance-administration-core/lib/constants'
 import { socialInsuranceAdministrationMessage } from '@island.is/application/templates/social-insurance-administration-core/lib/messages'
-import { getCurrencies } from '@island.is/application/templates/social-insurance-administration-core/lib/socialInsuranceAdministrationUtils'
+import {
+  getCurrencies,
+  getCategoriesOptions,
+  getTypesOptions,
+} from '@island.is/application/templates/social-insurance-administration-core/lib/socialInsuranceAdministrationUtils'
 import {
   Application,
   DefaultEvents,
@@ -23,23 +38,10 @@ import { formatCurrencyWithoutSuffix } from '@island.is/application/ui-component
 import { buildFormConclusionSection } from '@island.is/application/ui-forms'
 import isEmpty from 'lodash/isEmpty'
 import {
-  DIVIDENDS_IN_FOREIGN_BANKS,
-  FOREIGN_BASIC_PENSION,
-  FOREIGN_INCOME,
-  FOREIGN_PENSION,
-  INCOME,
-  INTEREST_ON_DEPOSITS_IN_FOREIGN_BANKS,
-  ISK,
-  RatioType,
-} from '../lib/constants'
-import {
   getApplicationAnswers,
   getApplicationExternalData,
-  getCategoriesOptions,
-  getTypesOptions,
 } from '../lib/incomePlanUtils'
 import { incomePlanFormMessage } from '../lib/messages'
-import { MONTHS } from '@island.is/application/templates/social-insurance-administration-core/lib/constants'
 
 export const IncomePlanForm: Form = buildForm({
   id: 'IncomePlanDraft',
@@ -49,7 +51,7 @@ export const IncomePlanForm: Form = buildForm({
   children: [
     buildSection({
       id: 'incomePlan',
-      title: incomePlanFormMessage.info.section,
+      title: socialInsuranceAdministrationMessage.incomePlan.subSectionTitle,
       children: [
         buildSubSection({
           id: 'incomePlanInstructions',
@@ -57,8 +59,12 @@ export const IncomePlanForm: Form = buildForm({
           children: [
             buildDescriptionField({
               id: 'instructions',
-              title: incomePlanFormMessage.info.instructionsTitle,
-              description: incomePlanFormMessage.info.instructionsDescription,
+              title:
+                socialInsuranceAdministrationMessage.incomePlanInstructions
+                  .title,
+              description:
+                socialInsuranceAdministrationMessage.incomePlanInstructions
+                  .instructions,
               space: 'containerGutter',
               doesNotRequireAnswer: false,
             }),
@@ -66,11 +72,13 @@ export const IncomePlanForm: Form = buildForm({
         }),
         buildSubSection({
           id: 'incomePlan',
-          title: incomePlanFormMessage.info.section,
+          title:
+            socialInsuranceAdministrationMessage.incomePlan.subSectionTitle,
           children: [
             buildTableRepeaterField({
               id: 'incomePlanTable',
-              title: incomePlanFormMessage.info.section,
+              title:
+                socialInsuranceAdministrationMessage.incomePlan.subSectionTitle,
               description: (application: Application) => {
                 const { incomePlanConditions, latestIncomePlan } =
                   getApplicationExternalData(application.externalData)
@@ -78,7 +86,7 @@ export const IncomePlanForm: Form = buildForm({
                 const baseMessage = hasLatestIncomePlan
                   ? incomePlanFormMessage.incomePlan
                       .currentIncomePlanDescription
-                  : incomePlanFormMessage.incomePlan.description
+                  : socialInsuranceAdministrationMessage.incomePlan.description
 
                 return {
                   ...baseMessage,
@@ -89,38 +97,52 @@ export const IncomePlanForm: Form = buildForm({
                   },
                 }
               },
-              formTitle: incomePlanFormMessage.incomePlan.registerIncome,
-              addItemButtonText: incomePlanFormMessage.incomePlan.addIncome,
-              saveItemButtonText: incomePlanFormMessage.incomePlan.saveIncome,
+              formTitle:
+                socialInsuranceAdministrationMessage.incomePlan.registerIncome,
+              addItemButtonText:
+                socialInsuranceAdministrationMessage.incomePlan.addIncome,
+              saveItemButtonText:
+                socialInsuranceAdministrationMessage.incomePlan.saveIncome,
               editField: true,
               editButtonTooltipText:
-                incomePlanFormMessage.incomePlan.editIncome,
+                socialInsuranceAdministrationMessage.incomePlan.editIncome,
               removeButtonTooltipText:
-                incomePlanFormMessage.incomePlan.removeIncome,
+                socialInsuranceAdministrationMessage.incomePlan.removeIncome,
               fields: {
                 incomeCategory: {
                   component: 'select',
-                  label: incomePlanFormMessage.incomePlan.incomeCategory,
+                  label:
+                    socialInsuranceAdministrationMessage.incomePlan
+                      .incomeCategory,
                   placeholder:
-                    incomePlanFormMessage.incomePlan.selectIncomeCategory,
+                    socialInsuranceAdministrationMessage.incomePlan
+                      .selectIncomeCategory,
                   displayInTable: false,
                   width: 'half',
                   isSearchable: true,
                   options: (application) => {
-                    return getCategoriesOptions(application.externalData)
+                    const { categorizedIncomeTypes } =
+                      getApplicationExternalData(application.externalData)
+
+                    return getCategoriesOptions(categorizedIncomeTypes)
                   },
                 },
                 incomeType: {
                   component: 'select',
-                  label: incomePlanFormMessage.incomePlan.incomeType,
+                  label:
+                    socialInsuranceAdministrationMessage.incomePlan.incomeType,
                   placeholder:
-                    incomePlanFormMessage.incomePlan.selectIncomeType,
+                    socialInsuranceAdministrationMessage.incomePlan
+                      .selectIncomeType,
                   width: 'half',
                   isSearchable: true,
                   updateValueObj: {
                     valueModifier: (application, activeField) => {
+                      const { categorizedIncomeTypes } =
+                        getApplicationExternalData(application.externalData)
+
                       const options = getTypesOptions(
-                        application.externalData,
+                        categorizedIncomeTypes,
                         activeField?.incomeCategory,
                       )
                       const selectedOption = options.find(
@@ -131,16 +153,22 @@ export const IncomePlanForm: Form = buildForm({
                     watchValues: 'incomeCategory',
                   },
                   options: (application, activeField) => {
+                    const { categorizedIncomeTypes } =
+                      getApplicationExternalData(application.externalData)
+
                     return getTypesOptions(
-                      application.externalData,
+                      categorizedIncomeTypes,
                       activeField?.incomeCategory,
                     )
                   },
                 },
                 currency: {
                   component: 'select',
-                  label: incomePlanFormMessage.incomePlan.currency,
-                  placeholder: incomePlanFormMessage.incomePlan.selectCurrency,
+                  label:
+                    socialInsuranceAdministrationMessage.incomePlan.currency,
+                  placeholder:
+                    socialInsuranceAdministrationMessage.incomePlan
+                      .selectCurrency,
                   isSearchable: true,
                   updateValueObj: {
                     valueModifier: (_, activeField) => {
@@ -183,18 +211,23 @@ export const IncomePlanForm: Form = buildForm({
                   options: [
                     {
                       value: RatioType.YEARLY,
-                      label: incomePlanFormMessage.incomePlan.annualIncome,
+                      label:
+                        socialInsuranceAdministrationMessage.incomePlan
+                          .annualIncome,
                     },
                     {
                       value: RatioType.MONTHLY,
-                      label: incomePlanFormMessage.incomePlan.monthlyIncome,
+                      label:
+                        socialInsuranceAdministrationMessage.incomePlan
+                          .monthlyIncome,
                     },
                   ],
                 },
                 equalForeignIncomePerMonth: {
                   component: 'input',
                   label:
-                    incomePlanFormMessage.incomePlan.equalForeignIncomePerMonth,
+                    socialInsuranceAdministrationMessage.incomePlan
+                      .equalForeignIncomePerMonth,
                   width: 'half',
                   type: 'number',
                   displayInTable: false,
@@ -235,7 +268,9 @@ export const IncomePlanForm: Form = buildForm({
                 },
                 equalIncomePerMonth: {
                   component: 'input',
-                  label: incomePlanFormMessage.incomePlan.equalIncomePerMonth,
+                  label:
+                    socialInsuranceAdministrationMessage.incomePlan
+                      .equalIncomePerMonth,
                   width: 'half',
                   type: 'number',
                   displayInTable: false,
@@ -276,7 +311,9 @@ export const IncomePlanForm: Form = buildForm({
                 },
                 incomePerYear: {
                   component: 'input',
-                  label: incomePlanFormMessage.incomePlan.incomePerYear,
+                  label:
+                    socialInsuranceAdministrationMessage.incomePlan
+                      .incomePerYear,
                   width: 'half',
                   type: 'number',
                   currency: true,
@@ -378,10 +415,10 @@ export const IncomePlanForm: Form = buildForm({
                     {
                       value: YES,
                       label:
-                        incomePlanFormMessage.incomePlan
+                        socialInsuranceAdministrationMessage.incomePlan
                           .monthlyDistributionOfIncome,
                       tooltip:
-                        incomePlanFormMessage.incomePlan
+                        socialInsuranceAdministrationMessage.incomePlan
                           .monthlyDistributionOfIncomeTooltip,
                     },
                   ],
@@ -605,9 +642,9 @@ export const IncomePlanForm: Form = buildForm({
                     value && formatCurrencyWithoutSuffix(value),
                 },
                 header: [
-                  incomePlanFormMessage.incomePlan.incomeType,
-                  incomePlanFormMessage.incomePlan.incomePerYear,
-                  incomePlanFormMessage.incomePlan.currency,
+                  socialInsuranceAdministrationMessage.incomePlan.incomeType,
+                  socialInsuranceAdministrationMessage.incomePlan.incomePerYear,
+                  socialInsuranceAdministrationMessage.incomePlan.currency,
                 ],
                 rows: ['incomeType', 'incomePerYear', 'currency'],
               },
