+++ libs/application/templates/social-insurance-administration/income-plan/src/lib/IncomePlanTemplate.spec.ts
@@ -1,4 +1,5 @@
 import { ApplicationTemplateHelper, YES } from '@island.is/application/core'
+import { RatioType } from '@island.is/application/templates/social-insurance-administration-core/lib/constants'
 import {
   Application,
   ApplicationStatus,
@@ -8,7 +9,6 @@ import {
   FormValue,
 } from '@island.is/application/types'
 import IncomePlanTemplate from './IncomePlanTemplate'
-import { RatioType } from './constants'
 
 const buildApplication = (data: {
   answers?: FormValue
