+++ libs/application/templates/social-insurance-administration/income-plan/src/lib/constants.ts
@@ -1,16 +1,2 @@
-export const FOREIGN_BASIC_PENSION = 'Erlendur grunnlífeyrir'
-export const FOREIGN_PENSION = 'Erlendur lífeyrir'
-export const FOREIGN_INCOME = 'Erlendar tekjur'
-export const INTEREST_ON_DEPOSITS_IN_FOREIGN_BANKS =
-  'Vextir af innstæðum í erlendum bönkum'
-export const DIVIDENDS_IN_FOREIGN_BANKS =
-  'Arður af hlutabréfa eign í erlendum bönkum'
-export const ISK = 'IKR'
-export const INCOME = 'Atvinnutekjur'
 export const INCOME_PLANS_CLOSED = 'INCOME_PLANS_CLOSED'
 export const NO_ACTIVE_APPLICATIONS = 'NO_ACTIVE_APPLICATIONS'
-
-export enum RatioType {
-  YEARLY = 'yearly',
-  MONTHLY = 'monthly',
-}
