+++ libs/application/templates/social-insurance-administration/income-plan/src/lib/incomePlanUtils.ts
@@ -1,14 +1,13 @@
 import { getValueViaPath } from '@island.is/application/core'
-import { Application, ExternalData } from '@island.is/application/types'
 import {
   CategorizedIncomeTypes,
   Eligible,
   IncomePlanConditions,
   IncomePlanRow,
-  LatestIncomePlan,
-  WithholdingTax,
-} from '../types'
-import { NO_ACTIVE_APPLICATIONS, INCOME_PLANS_CLOSED } from './constants'
+} from '@island.is/application/templates/social-insurance-administration-core/types'
+import { Application, ExternalData } from '@island.is/application/types'
+import { LatestIncomePlan, WithholdingTax } from '../types'
+import { INCOME_PLANS_CLOSED, NO_ACTIVE_APPLICATIONS } from './constants'
 import { incomePlanFormMessage } from './messages'
 
 export const getApplicationExternalData = (
@@ -98,48 +97,6 @@ export const getApplicationAnswers = (answers: Application['answers']) => {
   }
 }
 
-export const getOneInstanceOfCategory = (
-  categories: CategorizedIncomeTypes[],
-) => {
-  return [
-    ...new Map(
-      categories.map((category) => [category.categoryName, category]),
-    ).values(),
-  ]
-}
-
-export const getCategoriesOptions = (externalData: ExternalData) => {
-  const { categorizedIncomeTypes } = getApplicationExternalData(externalData)
-  const categories = getOneInstanceOfCategory(categorizedIncomeTypes)
-
-  return (
-    categories &&
-    categories.map((item) => {
-      return {
-        value: item.categoryName || '',
-        label: item.categoryName || '',
-      }
-    })
-  )
-}
-
-export const getTypesOptions = (
-  externalData: ExternalData,
-  categoryName: string | undefined,
-) => {
-  const { categorizedIncomeTypes } = getApplicationExternalData(externalData)
-  if (categoryName === undefined) return []
-
-  return categorizedIncomeTypes
-    .filter((item) => item.categoryName === categoryName)
-    .map((item) => {
-      return {
-        value: item.incomeTypeName || '',
-        label: item.incomeTypeName || '',
-      }
-    })
-}
-
 export const isEligible = (externalData: ExternalData): boolean => {
   const { isEligible } = getApplicationExternalData(externalData)
 
@@ -154,41 +111,3 @@ export const eligibleText = (externalData: ExternalData) => {
     ? incomePlanFormMessage.pre.isNotEligibleNoActiveApplicationDescription
     : incomePlanFormMessage.pre.isNotEligibleDescription
 }
-
-export const defaultIncomeTypes = [
-  {
-    income: 'yearly',
-    currency: 'IKR',
-    incomeType: 'Lífeyrissjóður',
-    incomePerYear: '0',
-    incomeCategory: 'Lífeyrissjóðstekjur',
-  },
-  {
-    income: 'yearly',
-    currency: 'IKR',
-    incomeType: 'Laun',
-    incomePerYear: '0',
-    incomeCategory: 'Atvinnutekjur',
-  },
-  {
-    income: 'yearly',
-    currency: 'IKR',
-    incomeType: 'Vextir af innistæðum',
-    incomePerYear: '0',
-    incomeCategory: 'Fjármagnstekjur',
-  },
-  {
-    income: 'yearly',
-    currency: 'EUR',
-    incomeType: 'Erlendur lífeyrir',
-    incomePerYear: '0',
-    incomeCategory: 'Lífeyrissjóðstekjur',
-  },
-  {
-    income: 'yearly',
-    currency: 'IKR',
-    incomeType: 'Vextir af verðbréfum',
-    incomePerYear: '0',
-    incomeCategory: 'Fjármagnstekjur',
-  },
-]
