+++ libs/application/templates/social-insurance-administration/income-plan/src/types.ts
@@ -1,14 +1,3 @@
-import { RatioType } from './lib/constants'
-
-export interface CategorizedIncomeTypes {
-  categoryCode?: string | null
-  categoryName?: string | null
-  categoryNumber?: number
-  incomeTypeCode?: string | null
-  incomeTypeName?: string | null
-  incomeTypeNumber?: number
-}
-
 export interface WithholdingTax {
   year?: number
   incomeTypes?: Array<WithholdingTaxIncomeType> | null
@@ -36,33 +25,6 @@ interface WithholdingTaxIncomeType {
   total?: number | null
 }
 
-export interface IncomePlanRow {
-  incomeType: string
-  incomeTypeNumber: number
-  incomeTypeCode: string
-  currency: string
-  incomeCategory: string
-  incomeCategoryNumber: number
-  incomeCategoryCode: string
-  income: RatioType
-  equalForeignIncomePerMonth?: string
-  equalIncomePerMonth?: string
-  incomePerYear: string
-  unevenIncomePerYear?: string
-  january?: string
-  february?: string
-  march?: string
-  april?: string
-  may?: string
-  june?: string
-  july?: string
-  august?: string
-  september?: string
-  october?: string
-  november?: string
-  december?: string
-}
-
 export interface LatestIncomePlan {
   year: number
   origin: string
@@ -80,15 +42,4 @@ interface IncomeType {
   incomeCategoryCode: string
   incomeCategoryName: string
   incomeCategoryNumber: number
-}
-
-export interface Eligible {
-  isEligible: boolean
-  reason: string
-  reasonCode: string
-}
-
-export interface IncomePlanConditions {
-  incomePlanYear: number
-  showTemporaryCalculations: boolean
 }
